CI: Remove cache dependencies job to ignore save gradle dependency cache
With this job present, every tag related build tries to save the
dependencies in a cache, but, this functionality has been broken for a
while, the 'example:build' task is constantly failing on CI (see below).

Example Build (1.23.1): https://buildkite.com/automattic/
android-dependency-catalog/builds/189
#01922323-d7e6-4526-82b5-f5dca9e6ed6a

FYI: No matter, we have decided to not use this project for this
purpose, as as such to have a single dependency cache for all projects,
and then restore this bigger than needed cache per project, only to use
a specific set of dependencies per project. Instead, we decided it is
better for each project to have a scheduled job, either weekly or daily,
which will trigger this 'save_gradle_dependency_cache' script and only
save the dependencies that it really needs.